[chore] Renamed the Game components
