+++ packages/dapp/game-app/src/components/Game/NotInGame.tsx
@@ -1,0 +1,14 @@
+
+
+
+
+
+export default function NotInGame() {
+    return (
+        <div className="not-in-game-container">
+            <h1>Not in Game</h1>
+            <p>You are not currently participating in any game.</p>
+            <p>Please join a game to start playing.</p>
+        </div>
+    );
+}
