+++ packages/dapp/game-app/src/components/Game/Payout/Payout.tsx
@@ -1,0 +1,34 @@
+
+
+/**
+ * @function Payout
+ * 
+ * @file ~/components/Game/Payout/Payout.tsx
+ * 
+ * @description This component shows the payout component
+ * that will be shown in a form of a modal or a separate page
+ * 
+ * It will be shown only to the winner(s) of the game
+ * after the whole game has successfully ended
+ * 
+ * the payout will exist only for the duration while the payout is being processed
+ * and otherwise it will not be shown
+ * 
+ * when the user is on the payout screen, keep it untill he refreshes, so basically
+ * that means that the payout will be displayed based on the state of the game
+ * 
+ * 
+ * @returns 
+ */
+
+
+
+export default function Payout() {
+    return (
+        <div className="payout-container">
+            <h1>Payout Component</h1>
+            <p>This is the payout component where you can implement your payout logic.</p>
+            {/* Add your payout UI and logic here */}
+        </div>
+    );
+}
