+++ packages/dapp/game-app/src/components/Game/index.ts
@@ -1,18 +1,5 @@
-// components/Game/index.ts
-export { default as GameMatch } from './GameMatch';
-export { default as MoveSelector } from './MoveSelector';
-export { default as OpponentCards } from './OpponentCards';
-export { default as Battlefield } from './Battlefield';
-export { default as MatchInfo } from './MatchInfo';
-export { default as GameDemo } from './GameDemo';
 
-// Re-export types for convenience
-export type {
-  Move,
-  GameState,
-  GameRound,
-  MatchData,
-  MoveOption,
-  RoundWinner,
-  MatchWinner,
-} from '@/types/game';
+export { default as Game } from './GamePage';
+export { default as GameInfo } from './GameInfo/GameInfo';
+export { default as OneOnOneInfo } from './GameInfo/OneOnOneInfo';
+export { default as TournamentInfo } from './GameInfo/TournamentInfo';
