--- packages/dapp/game-app/src/components/Game2/Match/MatchResult/MatchResult.tsx
@@ -1,17 +1,0 @@
-
-
-
-
-export default function MatchResult() {
-    
-    // this component will render the winner / loser animation in the content section
-    // of its parent match component
-
-    return (
-        <div className="match-result">
-            <h2>Match Result</h2>
-            <p>This section will display the result of the match, including the winner, final scores, and any other relevant information.</p>
-            {/* Add more detailed match result information here */}
-        </div>
-    );
-}
