--- packages/dapp/game-app/src/components/Game2/NotInGame.tsx
@@ -1,14 +1,0 @@
-
-
-
-
-
-export default function NotInGame() {
-    return (
-        <div className="not-in-game-container">
-            <h1>Not in Game</h1>
-            <p>You are not currently participating in any game.</p>
-            <p>Please join a game to start playing.</p>
-        </div>
-    );
-}
