--- packages/dapp/game-app/src/components/Game2/index.ts
@@ -1,5 +1,0 @@
-
-export { default as Game } from './GamePage';
-export { default as GameInfo } from './GameInfo/GameInfo';
-export { default as OneOnOneInfo } from './GameInfo/OneOnOneInfo';
-export { default as TournamentInfo } from './GameInfo/TournamentInfo';
