+++ packages/server/src/index.ts
@@ -61,5 +61,5 @@ app.listen(PORT, () => {
 })
 
 // This goes on and on and on
-// startWatchingGameRounds();
+startWatchingGameRounds();
